+++ src/resources/Tasks.ts
@@ -9,7 +9,13 @@ export default class Tasks {
   }
 
   create(params: CreateTaskParams): Promise<Task> {
+    const modifiedParams = {
+      ...params,
+      scheduled_for: params.scheduledFor?.toISOString() ?? params.scheduled_for,
+      scheduledFor: undefined,
+    };
+
     const defaultParams = { queue: "default" };
-    return this.client.post("tasks", { ...defaultParams, ...params });
+    return this.client.post("tasks", { ...defaultParams, ...modifiedParams });
   }
 }