+++ test/resources/Tasks.test.ts
@@ -26,4 +26,31 @@ describe("#create", () => {
       });
     });
   });
+
+  describe("with a scheduledFor param", () => {
+    test("makes a request to create a Task scheduled for the specified date", async () => {
+      const scheduledFor = new Date();
+      tasks.create({ request: { url: "" }, scheduledFor });
+      expect(client.post).toHaveBeenCalledWith("tasks", {
+        queue: "default",
+        request: { url: "" },
+        scheduled_for: scheduledFor.toISOString(),
+      });
+    });
+  });
+
+  describe("with a scheduledFor param and a scheduled_for param", () => {
+    test("makes a request, preferring scheduledFor param to deprecated scheduled_for param", async () => {
+      // eslint-disable-next-line @typescript-eslint/naming-convention
+      const scheduled_for = "2022-10-17T22:28:17.615Z";
+      const scheduledFor = new Date();
+      tasks.create({ request: { url: "" }, scheduledFor, scheduled_for });
+
+      expect(client.post).toHaveBeenCalledWith("tasks", {
+        queue: "default",
+        request: { url: "" },
+        scheduled_for: scheduledFor.toISOString(),
+      });
+    });
+  });
 });